Implement three-denomination currency: koku, bu, zeni
Replace single float koku field with three int fields matching
Rokugan's historical coinage:
  1 koku = 5 bu = 50 zeni
  1 bu = 10 zeni

Character.koku (int), Character.bu (int), Character.zeni (int)
replace the old Character.koku (float). Legacy float values are
auto-migrated in from_dict via _migrate_koku().

Updated across the codebase:
- character.py: New fields, total_zeni property, format_purse helper
- bot.py: Sheet display shows "Purse: X koku, Y bu, Z zeni"
- cog_inventory.py: Purse modal accepts koku/bu/zeni separately,
  validates against total balance in zeni, normalizes on save
- cog_edit.py: /edit field supports koku, bu, zeni individually
- cog_npc_builder.py: Wizard modal, state, form_to_state all
  handle three denominations
- npc_gen.py: Generates integer koku
- schools.py: Outfit parsing splits fractional koku into bu/zeni
- heritage.py: Koku grants as integer

+++ discord_bot/l5r_rules/npc_gen.py
@@ -118,7 +118,7 @@ def generate(
     # 4. Age from the Rank's range; koku = 1d10 x Rank (savings term only).
     age_lo, age_hi = RANK_AGE[rank]
     c.age = dice.rand_int_range(age_lo, age_hi)
-    c.koku = float(dice.roll_d10() * rank)
+    c.koku = dice.roll_d10() * rank
 
     # 5. Skills: caller-supplied school-skill names distributed per Rank. One
     #    becomes the specialty. No skill names are invented here.
